% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.env.R
\name{data.env}
\alias{data.env}
\title{dropbox.data}
\usage{
data.env(pkgname, Sys_dropbox_dir = "~/Dropbox/")
}
\arguments{
\item{pkgname}{The package name}

\item{Sys_dropbox_dir}{The location of the root dropbox folder Default: ~/Dropbox}
}
\description{
This file is used to load dropbox package data into the the global environment. It also identifies and returns all dropbox paths for the current package
}
\keyword{corelogic}
\keyword{import}
